shared/pyproject.toml
[project]
name = "shared"
version = "1.8.0001"
description = "Upbit Auto-Trading Platform shared lib (v1.8-0001)"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "pymysql>=1.1",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5"]

[tool.ruff]
line-length = 100
